Prep for redundant stream merging
I had a chat with Adrian last night where he clarified his original vision for RSM was that the stream relay would merge across *all* of the sources it had, rather than just the initial redundant routes computes for the egest, so that means merging ought to happen in the strem relay rather than the egest instance (which is the direction Steve and I were going in yesterday).

This change moves the egest forwarding from the receivers to a processor which receives from all of the sources in preparation for the merge code.